+++ modules/self-hosted/common.nix
@@ -1,7 +1,7 @@
 { config, lib, pkgs, ... }:
 
 let
-  dockerhub-secrets = config.sops.secrets."dockerhub-secrets".path;
+  dockerhub-secrets = config.ghostship.selfHostedSecrets.units."dockerhub-secrets".path;
   dockerhub-auth-script = pkgs.writeShellScriptBin "podman-dockerhub-auth" ''
     set -eu